# benefit-mankind/entity/src/main/java/cn/bdqn/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "微服务工程代码生成器,生成Dao、Service、Controller、Pojo、Feign、Swagger的JSON脚本",
    "version": "1.0.0",
    "title": "深圳黑马训练营代码生成器",
    "termsOfService": "https://github.com/shenkunlin/code-template.git",
    "contact": {
      "email": "shenkunlin@itcast.cn"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "petstore.swagger.io",
  "basePath": "/v2",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/bunk/{id}": {
      "delete": {
        "tags": [
          "BunkController"
        ],
        "summary": "根据ID删除Bunk",
        "description": "根据ID删除Bunk方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除Bunk方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除Bunk"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "BunkController"
        ],
        "summary": "根据ID修改Bunk",
        "description": "根据ID修改Bunk方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Bunk方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "Bunk",
            "description": "传入Bunk的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Bunk"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Bunk"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "BunkController"
        ],
        "summary": "根据ID查询Bunk",
        "description": "根据ID查询Bunk方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Bunk方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Bunk"
            ,
            "schema": {
              "$ref": "#/definitions/Result«Bunk»"
            }
          }
        }
      }
    },
    "/bunk/search": {
      "post": {
        "tags": [
          "BunkController"
        ],
        "summary": "不带分页条件搜索Bunk",
        "description": "不带分页条件搜索Bunk方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Bunk",
            "description": "传入Bunk的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Bunk"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Bunk"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Bunk»»"
            }
          }
        }
      }
    },
    "/bunk/search/{page}/{size}": {
      "get": {
        "tags": [
          "BunkController"
        ],
        "summary": "分页列表查询Bunk",
        "description": "分页列表查询Bunk方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Bunk"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Bunk»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "BunkController"
        ],
        "summary": "分页条件搜索Bunk",
        "description": "分页条件搜索Bunk方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "Bunk",
            "description": "传入Bunk的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Bunk"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索Bunk"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Bunk»»"
            }
          }
        }
      }
    },
...
    "/departments/{id}": {
      "delete": {
        "tags": [
          "DepartmentsController"
        ],
        "summary": "根据ID删除Departments",
        "description": "根据ID删除Departments方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除Departments方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除Departments"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "DepartmentsController"
        ],
        "summary": "根据ID修改Departments",
        "description": "根据ID修改Departments方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Departments方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "Departments",
            "description": "传入Departments的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Departments"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Departments"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "DepartmentsController"
        ],
        "summary": "根据ID查询Departments",
        "description": "根据ID查询Departments方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Departments方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Departments"
            ,
            "schema": {
              "$ref": "#/definitions/Result«Departments»"
            }
          }
        }
      }
    },
    "/departments/search": {
      "post": {
        "tags": [
          "DepartmentsController"
        ],
        "summary": "不带分页条件搜索Departments",
        "description": "不带分页条件搜索Departments方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Departments",
            "description": "传入Departments的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Departments"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Departments"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Departments»»"
            }
          }
        }
      }
    },
    "/departments/search/{page}/{size}": {
      "get": {
        "tags": [
          "DepartmentsController"
        ],
        "summary": "分页列表查询Departments",
        "description": "分页列表查询Departments方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Departments"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Departments»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "DepartmentsController"
        ],
        "summary": "分页条件搜索Departments",
        "description": "分页条件搜索Departments方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "Departments",
            "description": "传入Departments的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Departments"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索Departments"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Departments»»"
            }
          }
        }
      }
    },
...
    "/dept/{id}": {
      "delete": {
        "tags": [
          "DeptController"
        ],
        "summary": "根据ID删除Dept",
        "description": "根据ID删除Dept方法详情",
        "operationId": "deleteUsingDELETE",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID删除Dept方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID删除Dept"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "put": {
        "tags": [
          "DeptController"
        ],
        "summary": "根据ID修改Dept",
        "description": "根据ID修改Dept方法详情",
        "operationId": "updateUsingPUT",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Dept方法详情",
            "required": true
          },
          {
            "in": "body",
            "name": "Dept",
            "description": "传入Dept的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Dept"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Dept"
            ,
            "schema": {
              "$ref": "#/definitions/Result"
            }
          }
        }
      },
      "get": {
        "tags": [
          "DeptController"
        ],
        "summary": "根据ID查询Dept",
        "description": "根据ID查询Dept方法详情",
        "operationId": "findByIdUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "根据ID修改Dept方法详情",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "根据ID修改Dept"
            ,
            "schema": {
              "$ref": "#/definitions/Result«Dept»"
            }
          }
        }
      }
    },
    "/dept/search": {
      "post": {
        "tags": [
          "DeptController"
        ],
        "summary": "不带分页条件搜索Dept",
        "description": "不带分页条件搜索Dept方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Dept",
            "description": "传入Dept的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Dept"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Dept"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Dept»»"
            }
          }
        }
      }
    },
    "/dept/search/{page}/{size}": {
      "get": {
        "tags": [
          "DeptController"
        ],
        "summary": "分页列表查询Dept",
        "description": "分页列表查询Dept方法详情",
        "operationId": "findPageUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Dept"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Dept»»"
            }
          }
        }
      },
      "post": {
        "tags": [
          "DeptController"
        ],
        "summary": "分页条件搜索Dept",
        "description": "分页条件搜索Dept方法详情",
        "operationId": "findPageUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "page",
            "description": "当前页",
            "required": true
          },
          {
            "in": "path",
            "name": "size",
            "description": "每页显示条数",
            "required": true
          },
          {
            "in": "body",
            "name": "Dept",
            "description": "传入Dept的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Dept"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "分页条件搜索Dept"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Dept»»"
            }
          }
        }
      }
    },
...
    "/doctor/search": {
      "post": {
        "tags": [
          "DoctorController"
        ],
        "summary": "不带分页条件搜索Doctor",
        "description": "不带分页条件搜索Doctor方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Doctor",
            "description": "传入Doctor的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Doctor"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Doctor"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Doctor»»"
            }
          }
        }
      }
    },
...
    "/doctordiagnose/search": {
      "post": {
        "tags": [
          "DoctordiagnoseController"
        ],
        "summary": "不带分页条件搜索Doctordiagnose",
        "description": "不带分页条件搜索Doctordiagnose方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Doctordiagnose",
            "description": "传入Doctordiagnose的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Doctordiagnose"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Doctordiagnose"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Doctordiagnose»»"
            }
          }
        }
      }
    },
...
    "/druginfo/search": {
      "post": {
        "tags": [
          "DruginfoController"
        ],
        "summary": "不带分页条件搜索Druginfo",
        "description": "不带分页条件搜索Druginfo方法详情",
        "operationId": "findListUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "Druginfo",
            "description": "传入Druginfo的JSON对象",
            "required": false
              ,
              "schema": {
                "$ref": "#/definitions/Druginfo"
              }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid status value(无效的状态值)"
                      },
          "403": {
            "description": "403 Forbidden(请求被拒绝)"
                      },
          "404": {
            "description": "not found(没有找到相关资源)"
                      },
          "405": {
            "description": "Invalid input(无效的输入)"
                      },
          "500": {
            "description": "服务器内部错误"
                      },
          "200": {
            "description": "不带分页搜索Druginfo"
            ,
            "schema": {
              "$ref": "#/definitions/Result«List«Druginfo»»"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Result": {
      "type": "object",
      "required": [
        "code",
        "data",
        "flag",
        "message"
      ],
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32",
          "description": "返回状态码,20000:成功,20001:失败,20002:用户名或密码错误,20003:权限不足,20004:远程调用失败,20005:重复操作,20006:没有对应的数据"
        },
        "data": {
          "type": "object",
          "description": "逻辑数据"
        },
        "flag": {
          "type": "boolean",
          "example": false,
          "description": "执行是否成功,true:成功,false:失败"
        },
        "message": {
          "type": "string",
          "description": "提示信息"
        }
      },
      "description": "Result"
    },
...
    "Bunk": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "integer",
          "format": "int32",
          "description": ""
        },
        "BunkID": {
          "type": "integer",
          "format": "int32",
          "description": "床位号"
        },
        "RoomID": {
          "type": "integer",
          "format": "int32",
          "description": "房间表外键ID"
        },
        "BunkStatus": {
          "type": "integer",
          "format": "int32",
          "description": "是否住人（0，未住1，住）"
        }
      },
      "description": "Bunk"
    },
...
  },
...
}